Workaround for python pip failing silently on 64bit
The workaround is to install libuuid-devel.

http://goo.gl/EzfEGy
http://goo.gl/nrpeSV

There's a bug(?) in 64-bit Cygwin which causes ctypes.util to segfault
when trying to find libuuid (/usr/bin/cyguuid-1.dll). The fix is to
install libuuid-devel from Cygwin setup. I found this from an issue
filed against requests.py, but it's noted (and worked around in
different ways) in a few other places, too.